Updated code and files
Made several updates to the code:
I made the 64-bit database the default and moved the 32-bit database to a new directory; I deleted the additional sheets in the Excel workbook. So it went from three sheets to one; I added a constructor to the cSqlite demo file to create the test file anytime the object was instantiated. This fixed a bug within the code.